scripts/load/requirements.txt
azure-search-documents>=11.4.0
azure-identity>=1.17.0
python-dotenv>=1.0.0
